requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.35.0
blis==0.7.11
catalogue==2.0.10
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.1
confection==0.1.4
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.6
fsspec==2024.2.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.21.4
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openmeteo_requests==1.2.0
openmeteo_sdk==1.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.2
regex==2023.12.25
requests==2.31.0
requests-cache==1.2.0
retry==0.9.2
retry-requests==2.0.0
safetensors==0.4.2
sentencepiece==0.2.0
setuptools==69.1.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
thinc==8.2.3
tokenizers==0.15.2
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websockets==12.0
wheel==0.42.0